+++ android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">projectsecu1</string>
+    <string name="app_name">My KTM Tracker</string>
 </resources>
