+++ src/assets/fonts/index.js
@@ -1,0 +1,37 @@
+import { useCallback } from 'react';
+import { Text, View, StyleSheet } from 'react-native';
+import { useFonts } from 'expo-font';
+import * as SplashScreen from 'expo-splash-screen';
+
+SplashScreen.preventAutoHideAsync();
+
+export default function Fonts() {
+  const [fontsLoaded] = useFonts({
+    'Plus Jakarta Sans': require('./PlusJakartaSans-VariableFont_wght.ttf'),
+  });
+
+  const onLayoutRootView = useCallback(async () => {
+    if (fontsLoaded) {
+      await SplashScreen.hideAsync();
+    }
+  }, [fontsLoaded]);
+
+  if (!fontsLoaded) {
+    return null;
+  }
+
+  return (
+    <View style={styles.container} onLayout={onLayoutRootView}>
+      <Text style={{ fontFamily: 'Plus Jakarta Sans', fontSize: 30 }}>Inter Black</Text>
+      <Text style={{ fontSize: 30 }}>Platform Default</Text>
+    </View>
+  );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
