+++ src/style/color.js
@@ -1,0 +1,3 @@
+export const PRIMARY_COLOR = '#EA5455'
+export const SECONDARY_COLOR = '#EA5455'
+export const TERTIARY_COLOR = '#EA5455'
